scripts/requirements.txt
hypothesis==6.39.3
tqdm
pyyaml
